+++ packages/client/package.json
@@ -18,7 +18,7 @@
   },
   "dependencies": {
     "axios": "^1.7.7",
-    "clsx": "^2.0.0",
+    "clsx": "^2.1.1",
     "lucide-react": "^0.438.0",
     "next": "^14.2.7",
     "react": "^18.2.0",
